wined3d: Initialise WINED3DVERTEXELEMENT's Reg field in IWineD3DVertexDeclarationImpl_ParseDeclaration9.

+++ include/wine/wined3d_types.h
@@ -507,7 +507,7 @@ typedef struct _WINED3DVERTEXELEMENT {
   BYTE    Method;
   BYTE    Usage;
   BYTE    UsageIndex;
-  BYTE    Reg; /* DirectX 8 */
+  int     Reg; /* DirectX 8 */
 } WINED3DVERTEXELEMENT, *LPWINED3DVERTEXELEMENT;
